Dialer Lookup [2/2]
USA-RedDragon edits: Forward port to oreo.

This is a squash of 3 commits:

Enable reverse lookup by adding it as default PhoneNumberService

This is based on I66396b21cf4573e017bc2307441cde1959d34154 but puts
the ReverseLookupService directly into the default ObjectFactory
instead of fiddling with another ServiceFactory as in older versions.

Without this change the reverse lookup will never work because its
callback is not used anywhere.

String change for forward lookup

Change-Id: I9de5e115f631a885847f86ae7d7d86a2c89d29f2